creates IconImage composable in another file
