+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/templates/Template2.kt
@@ -1,20 +1,13 @@
 package com.revenuecat.purchases.ui.revenuecatui.templates
 
-import android.graphics.Bitmap
-import android.net.Uri
 import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
@@ -23,32 +16,23 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.core.graphics.drawable.toBitmap
-import com.revenuecat.purchases.paywalls.PaywallData
 import com.revenuecat.purchases.ui.revenuecatui.InternalPaywallView
 import com.revenuecat.purchases.ui.revenuecatui.UIConstant
 import com.revenuecat.purchases.ui.revenuecatui.composables.Footer
+import com.revenuecat.purchases.ui.revenuecatui.composables.IconImage
 import com.revenuecat.purchases.ui.revenuecatui.composables.PaywallBackground
 import com.revenuecat.purchases.ui.revenuecatui.composables.PurchaseButton
-import com.revenuecat.purchases.ui.revenuecatui.composables.RemoteImage
 import com.revenuecat.purchases.ui.revenuecatui.data.PaywallViewModel
 import com.revenuecat.purchases.ui.revenuecatui.data.PaywallViewState
 import com.revenuecat.purchases.ui.revenuecatui.data.TestData
 import com.revenuecat.purchases.ui.revenuecatui.data.processed.TemplateConfiguration
 import com.revenuecat.purchases.ui.revenuecatui.data.selectedLocalization
-import com.revenuecat.purchases.ui.revenuecatui.extensions.defaultAppIconPlaceholder
-import com.revenuecat.purchases.ui.revenuecatui.helpers.isInPreviewMode
 
 private object Template2UIConstants {
     val maxIconWidth = 140.dp
@@ -82,7 +66,11 @@ private fun Template2MainContent(state: PaywallViewState.Loaded, viewModel: Payw
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.spacedBy(UIConstant.defaultVerticalSpacing, Alignment.CenterVertically),
     ) {
-        IconImage(state.templateConfiguration.images.iconUri)
+        IconImage(
+            uri = state.templateConfiguration.images.iconUri,
+            maxWidth = Template2UIConstants.maxIconWidth,
+            iconCornerRadius = Template2UIConstants.iconCornerRadius,
+        )
         val localizedConfig = state.selectedLocalization
         val colors = state.templateConfiguration.getCurrentColors()
         Text(
@@ -102,55 +90,6 @@ private fun Template2MainContent(state: PaywallViewState.Loaded, viewModel: Payw
         state.templateConfiguration.packages.all.forEach { packageInfo ->
             SelectPackageButton(state, packageInfo, viewModel)
         }
-    }
-}
-
-@Composable
-private fun IconImage(uri: Uri?) {
-    uri?.let {
-        Column(modifier = Modifier.widthIn(max = Template2UIConstants.maxIconWidth)) {
-            val modifier = Modifier
-                .aspectRatio(ratio = 1f)
-                .widthIn(max = Template2UIConstants.maxIconWidth)
-                .clip(RoundedCornerShape(Template2UIConstants.iconCornerRadius))
-            if (uri.toString().contains(PaywallData.defaultAppIconPlaceholder)) {
-                AppIcon(
-                    modifier = modifier,
-                )
-            } else {
-                RemoteImage(
-                    urlString = uri.toString(),
-                    modifier = modifier,
-                    contentScale = ContentScale.Crop,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun AppIcon(
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-
-    if (isInPreviewMode()) {
-        Box(
-            modifier = Modifier
-                .background(color = MaterialTheme.colorScheme.primary)
-                .size(100.dp),
-        )
-    } else {
-        val appIconResId = remember {
-            val packageManager = context.packageManager
-            context.applicationInfo.loadIcon(packageManager)
-        }
-        Image(
-            bitmap = appIconResId.toBitmap(config = Bitmap.Config.ARGB_8888).asImageBitmap(),
-            contentDescription = null,
-            modifier = modifier,
-            contentScale = ContentScale.Crop,
-        )
     }
 }
